# api.yaml
openapi: 3.0.4
info:
  title: Bilheteira AFC Coimbra
  description: |-
    API para a venda de bilhetes para a Divisão de Elite da Associação de Futebol de Coimbra
  contact:
    email: afcoimbra@afcoimbra.com
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Servidor local
...
security:
  - bearerAuth: []
paths:
...
  /equipa:
    get:
      tags: [Equipa]
      summary: Todas as equipas
      operationId: getAllEquipas
      responses:
        '200':
          description: Lista de equipas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipa'
        '500':
          description: Falha no servidor
    post:
      tags: [Equipa]
      summary: Cria uma nova equipa
      operationId: createEquipa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipa'
      responses:
        '201':
          description: Equipa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipa'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /equipa/{id_equipa}:
    parameters:
      - name: id_equipa
        in: path
        required: true
        description: ID da equipa
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Equipa]
      summary: Encontrar equipa pelo ID
      operationId: getEquipaById
      responses:
        '200':
          description: Equipa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipa'
        '404':
          description: Equipa não encontrada
        '500':
          description: Falha no servidor
    put:
      tags: [Equipa]
      summary: Atualiza uma equipa pelo ID
      operationId: updateEquipa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipa'
      responses:
        '200':
          description: Equipa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipa'
        '400':
          description: Dados inválidos
        '404':
          description: Equipa não encontrada
        '500':
          description: Falha no servidor
    delete:
      tags: [Equipa]
      summary: Apaga uma equipa pelo ID
      operationId: deleteEquipa
      responses:
        '204':
          description: Equipa eliminada com sucesso
        '404':
          description: Equipa não encontrada
        '500':
          description: Falha no servidor
  /estadio:
    get:
      tags: [Estádio]
      summary: Todos os estádios
      operationId: getAllEstadios
      responses:
        '200':
          description: Lista de estádios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Estadio'
        '500':
          description: Falha no servidor
    post:
      tags: [Estádio]
      summary: Cria um novo estádio
      operationId: createEstadio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estadio'
      responses:
        '201':
          description: Estádio criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estadio'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /estadio/{id_estadio}:
    parameters:
      - name: id_estadio
        in: path
        required: true
        description: ID do estádio
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Estádio]
      summary: Encontrar estádio pelo ID
      operationId: getEstadioById
      responses:
        '200':
          description: Estádio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estadio'
        '404':
          description: Estádio não encontrado
        '500':
          description: Falha no servidor
    put:
      tags: [Estádio]
      summary: Atualiza um estádio pelo ID
      operationId: updateEstadio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estadio'
      responses:
        '200':
          description: Estádio atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estadio'
        '400':
          description: Dados inválidos
        '404':
          description: Estádio não encontrado
        '500':
          description: Falha no servidor
    delete:
      tags: [Estádio]
      summary: Apaga um estádio pelo ID
      operationId: deleteEstadio
      responses:
        '204':
          description: Estádio eliminado com sucesso
        '404':
          description: Estádio não encontrado
        '500':
          description: Falha no servidor
  /jogo:
    get:
      tags: [Jogo]
      summary: Todos os jogos
      operationId: getAllJogos
      parameters:
        - name: jornada
          in: query
          required: false
          description: Filtrar jogos por número de jornada
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Lista de jogos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Jogo'
        '500':
          description: Falha no servidor
    post:
      tags: [Jogo]
      summary: Cria um novo jogo
      operationId: createJogo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jogo'
      responses:
        '201':
          description: Jogo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jogo'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /jogo/{id_jogo}:
    parameters:
      - name: id_jogo
        in: path
        required: true
        description: ID do jogo
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Jogo]
      summary: Encontrar jogo pelo ID
      operationId: getJogoById
      responses:
        '200':
          description: Jogo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jogo'
        '404':
          description: Jogo não encontrado
        '500':
          description: Falha no servidor
    put:
      tags: [Jogo]
      summary: Atualiza um jogo pelo ID
      operationId: updateJogo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jogo'
      responses:
        '200':
          description: Jogo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jogo'
        '400':
          description: Dados inválidos
        '404':
          description: Jogo não encontrado
        '500':
          description: Falha no servidor
    delete:
      tags: [Jogo]
      summary: Apaga um jogo pelo ID
      operationId: deleteJogo
      responses:
        '204':
          description: Jogo eliminado com sucesso
        '404':
          description: Jogo não encontrado
        '500':
          description: Falha no servidor
...
components:
  schemas:
...
    Equipa:
      type: object
      required:
        - nome
      properties:
        id_equipa:
          type: integer
          format: int64
          description: Chave primária da entidade Equipa
          example: 1
        nome:
          type: string
          description: Nome completo do Clube
          example: Vigor da Mocidade
        fundacao:
          type: string
          format: date
          description: Data de fundação do Clube
          example: "1999-09-09"
    Estadio:
      type: object
      required:
        - nome
        - morada
        - lotacao
        - preco_base
      properties:
        id_estadio:
          type: integer
          format: int64
          description: Chave primária da entidade Estádio
          example: 1
        nome:
          type: string
          description: Nome completo do Estádio
          example: Campo dos Sardões
        morada:
          type: string
          description: Morada do Estádio
          example: Rua da Rua, Coimbra
        lotacao:
          type: integer
          format: int64
          description: Lotação do Estádio
          example: 2000
        preco_base:
          type: number
          format: float
          description: Preço base do Estádio
          example: 4.50
        id_equipa:
          type: integer
          format: int64
          description: ID da equipa associada ao estádio
          example: 1
    Jogo:
      type: object
      required:
        - data
        - hora
        - jornada
        - id_estadio
        - id_equipa_casa
        - id_equipa_fora
      properties:
        id_jogo:
          type: integer
          format: int64
          description: Chave primária da entidade Jogo
          example: 1
        data:
          type: string
          format: date
          description: Data do jogo
          example: "2025-06-14"
        hora:
          type: string
          description: Hora do jogo (formato HH:MM)
          example: "15:00"
        jornada:
          type: integer
          format: int64
          description: Jornada a que o jogo pertence
          example: 10
        id_estadio:
          type: integer
          format: int64
          description: ID do estádio onde o jogo se realiza
          example: 1
        id_equipa_casa:
          type: integer
          format: int64
          description: ID da equipa que joga em casa
          example: 2
        id_equipa_fora:
          type: integer
          format: int64
          description: ID da equipa que joga fora
          example: 3
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT